mysite/requirements1.txt
dj-database-url==0.5.0
Django==2.2.11
django-classy-tags==1.0.0
django-cms==3.7.1
django-filer==1.7.0
django-formtools==2.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
django-sekizai==1.1.0
django-treebeard==4.3.1
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.2.0
djangocms-bootstrap4==1.5.0
djangocms-file==2.4.0
djangocms-googlemap==1.3.0
djangocms-icon==1.5.0
djangocms-installer==1.2.1
djangocms-link==2.5.0
djangocms-picture==2.3.0
djangocms-snippet==2.2.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.8.0
djangocms-video==2.2.0
easy-thumbnails==2.7
html5lib==1.0.1
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
tzlocal==2.0.0
Unidecode==1.1.1
webencodings==0.5.1
